première partie finie et débuguée

+++ ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: lgaudet- <lgaudet-@student.42lyon.f>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 16:17:49 by lgaudet-          #+#    #+#             */
-/*   Updated: 2020/11/23 16:17:55 by lgaudet-         ###   ########lyon.fr   */
+/*   Updated: 2020/11/24 14:28:12 by lgaudet-         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
 void	*ft_bzero(void *pointer, size_t count)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (count <= 0)
