+++ ft_memchr.c
@@ -6,18 +6,18 @@
 /*   By: lgaudet- <lgaudet-@student.42lyon.f>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 16:19:03 by lgaudet-          #+#    #+#             */
-/*   Updated: 2020/11/23 16:19:07 by lgaudet-         ###   ########lyon.fr   */
+/*   Updated: 2020/11/24 14:37:43 by lgaudet-         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
 void	*ft_memchr(const void *block, int ch, size_t size)
 {
 	char *b;
 
 	b = (void*)block;
-	while ((void*)b - block < size)
+	while ((size_t)((void*)b - block) < size)
 		if (*b == ch)
 			return (b);
 		else
