+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: lgaudet- <lgaudet-@student.42lyon.f>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 16:20:10 by lgaudet-          #+#    #+#             */
-/*   Updated: 2020/11/23 16:20:13 by lgaudet-         ###   ########lyon.fr   */
+/*   Updated: 2020/11/24 14:38:41 by lgaudet-         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
 void	*ft_memset(void *pointer, int value, size_t count)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (i < count)
